Version 2.0: Rewriting of code.
I decided to go back to this "old" project and rewrite some of the spaghetti code I created. Of course, most of it didn't require more than a few space additions, but some of them are changed a bit.
SlapNR now has less checks and relies on ULX's minimum and maximum for damage.
ReloadMap now has a timer so it waits around 1.5 seconds before it reloads/restarts the map.